meibot/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
apify-client==1.0.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coala-utils==0.5.1
colorama==0.4.6
discord.py==2.2.2
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
google-api-core==2.11.0
google-api-python-client==2.85.0
googleapis-common-protos==1.59.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.4
packaging==23.1
pefile==2023.2.7
Pillow==9.5.0
pip==23.1.2
playwright==1.32.1
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==9.0.4
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pyparsing==3.0.9
pytesseract==0.3.10
python-dotenv==1.0.0
pywin32-ctypes==0.2.0
PyYAML==6.0
regex==2023.5.5
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
sentence-transformers==2.2.2
setuptools==67.7.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
twitterscraper==1.6.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wheel==0.40.0
yarl==1.8.2
